// src/interfaces/ruleBuilderInterface.ts
export interface ruleBuilderAttributes {
  id?: string;
  rule_code?: string;
  status?: string;
  rule_name?: string;
  module_id?: string;
  program_id: string;
  module_name?: string;
  module_code?: string;
  rule_type?: string;
  effective_start_date?: string;
  effective_end_date?: string;
  created_on?: string;
  created_by: string;
  updated_by: string;
  is_enabled?: boolean;
  modified_by?: string;
  modified_on?: string;
  rule_event_id?: string;
  is_deleted?: boolean;
  page?: string;
  limit?: string;
}